requirements.txt
python-telegram-bot==20.7
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
